pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rideshala"
version = "0.1.0"
description = "Open-source AI-powered motorcycle comparison for Indian riders"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "RideShala Contributors"},
]

dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    # Vector database
    "qdrant-client>=1.12",
    # Cache
    "redis[hiredis]>=5.2",
    # Search
    "meilisearch>=0.31",
    # LLM providers
    "openai>=1.60",
    "anthropic>=0.40",
    # AI orchestration
    "langgraph>=0.2",
    "langchain-core>=0.3",
    # Embeddings & reranking
    "sentence-transformers>=3.3",
    # Resilience
    # pybreaker removed — uses Tornado @gen.coroutine, broken on Python 3.13+
    # Circuit breaker is built-in at packages/ai/llm/circuit_breaker.py
    # Observability
    "structlog>=24.4",
    "prometheus-client>=0.21",
    # Security
    "bleach>=6.2",
    "pydantic>=2.10",
    "PyJWT>=2.9",
    # HTTP client
    "httpx>=0.28",
    # Utilities
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "pip-audit>=2.7",
    "httpx>=0.28",  # for TestClient
]
eval = [
    "ragas>=0.2",
]
gpu = [
    # vLLM is installed separately on GPU machines
    # "vllm>=0.6",
]

[tool.setuptools.packages.find]
include = ["packages*", "services*"]

[tool.setuptools.package-dir]
# Allow `from packages.ai...` and `from services.api...` imports
"" = "."

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.lint.isort]
known-first-party = ["packages", "services", "app"]

[tool.pytest.ini_options]
testpaths = ["services/api/tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["services/api/app", "packages/ai"]
branch = true
omit = ["*/tests/*", "*/alembic/*", "*/migrations/*"]

[tool.coverage.report]
fail_under = 40
show_missing = true
skip_empty = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Relaxed for MVP
